Bulk edit bookmarks (#32)
Savory has allowed you to select multiple bookmarks for a while, but the only extra affordance it offered was to delete them. With this change, we allow users to add or remove tags to multiple bookmarks at once.

Once you select 1 or more bookmarks, a new UI pops up in the sidebar which shows you a set union of all the tags for selected bookmarks. You can delete a tag from this UI and it will be pruned from all the bookmarks in the selection that have that tag.

Similarly, you can add a new tag via this UI and all the bookmarks in selection will get it.