Use jest for testing (#38)
## We have tests

I know, I know this project has been under development for more than 3 years and I am only now adding tests. Better late than never!

## Never a dull day in JavaScript

This commit sets up `jest` in the project. It took surprisingly longer than expected. I hate to repeat the "javascript is hard" meme but it exists for a reason. I knew all the pieces so was able to piece them together but it took 2hrs more than it should have taken to integrate the most popular javascript testing framework in a straightforward vuejs app.

The complexity was due to jest test runner being an alternate `webpack` in a way. That's why I had to fork the babel config to use different configuration depending on the environment—browser or tests.

## Feature alert: Tag autocomplete by frequency

Anyway, the first test (and the only test for now) is for a bugfix (or feature depending on how you hold it). We are improving tag suggestions by suggesting the more frequent tag over other less frequently used tags.

For example, if you have 100 bookmarks tagged with "systems" and only a few tagged with "systemd", we would still suggest "systemd" if you typed "sys" even though it is much more likely that you want "systems".

It's fixed now!