Extend backend client to do exponential backoff retries on timeouts (#42)
Backend API requests sometime time out. This PR implements retries to handle timeout errors on the client.

Not all requests are safe to retry -- so we have introduced an `urlsToRetry` config option to opt-in.

This PR opts-in the "create new bookmark" API operation (`/bookmarks/add`). It is probably NOT safe to retry but let's collect some data before making it idempotent.

Exponential backoff code adapted from [vercel/swr][swr].

[swr]: https://github.com/vercel/swr/blob/3bbbf5b86be0868963c12906816a7bc1ded84c15/src/utils/config.ts#L29